+++ linode/resource_linode_stackscript.go
@@ -199,27 +199,7 @@ func resourceLinodeStackscriptRead(d *schema.ResourceData, meta interface{}) err
 		return err
 	}
 
-	if stackscript.UserDefinedFields == nil {
-		if err := d.Set("user_defined_fields", nil); err != nil {
-			return fmt.Errorf("Error setting user_defined_fields: %s", err)
-		}
-	} else {
-		var udfs []map[string]string
-		for _, udf := range *stackscript.UserDefinedFields {
-			udfs = append(udfs, map[string]string{
-				"default": udf.Default,
-				"example": udf.Example,
-				"many_of": udf.ManyOf,
-				"one_of":  udf.OneOf,
-				"label":   udf.Label,
-				"name":    udf.Name,
-			})
-		}
-		if err := d.Set("user_defined_fields", udfs); err != nil {
-			return fmt.Errorf("Error setting user_defined_fields: %s", err)
-		}
-	}
-
+	setStackScriptUserDefinedFields(d, stackscript)
 	return nil
 }